+++ examples/static-hbase/catS3file.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+file=$1
+cd /home/hadoop/contents
+cat $file