+++ examples/static-hbase/outputs.tf
@@ -9,3 +9,11 @@ output "rootdir-bucket" {
 output "cluster" {
   value = module.emr-hbase
 }
+
+output "ec2-key" {
+  value = module.emr_key_pair
+}
+
+output "private-key" {
+  value = tls_private_key.emr_private_key.private_key_pem
+}