+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # TAMR AWS EMR module
 
+## v0.13.0 - January 12th 2021
+* Adds optional input `bootstrap_actions` to run ordered list of actions on cluster nodes before Hadoop starts.
+* Adds example usage of `bootstrap_actions` in examples/static-hbase
+
 ## v0.12.0 - January 8th 2021
 * Exposes variable json_configuration_bucket_key to make the path to the EMR configuration file
  settable.